refactor: Move more modals to their own component
