+++ app/javascript/src/components/editor/EditorActions.svelte
@@ -1,8 +1,7 @@
 <script>
   import { fly } from "svelte/transition"
-  import { currentProject, isSignedIn, isMobile, currentProjectUUID } from "../../stores/editor"
-  import ScriptImporter from "./ScriptImporter.svelte"
-  import TranslationKeys from "./TranslationKeys/TranslationKeys.svelte"
+  import { currentProject, isSignedIn, isMobile, currentProjectUUID, modal } from "../../stores/editor"
+  import { Modal } from "../../constants/Modal"
   import Compiler from "./Compiler.svelte"
   import Settings from "./Settings.svelte"
   import Shortcuts from "./Shortcuts.svelte"
@@ -19,10 +18,15 @@
   {#if !$isMobile}
     <Shortcuts />
     <Settings />
-    <TranslationKeys />
+
+    <button class="button button--secondary button--square" on:click={() => modal.show(Modal.TranslationKeys)}>
+      Translations
+    </button>
 
     {#if $isSignedIn && $currentProject?.is_owner}
-      <ScriptImporter />
+      <button class="button button--secondary button--square" on:click={() => modal.show(Modal.ScriptImporter)}>
+        Import Script
+      </button>
     {/if}
 
     <Compiler />
