+++ app/javascript/src/components/editor/Modals/Modals.svelte
@@ -3,6 +3,8 @@
   import CreateProjectModal from "./CreateProjectModal.svelte"
   import LineFinderModal from "./LineFinderModal.svelte"
   import ItemFinderModal from "./ItemFinderModal.svelte"
+  import TranslationKeysModal from "./TranslationKeysModal.svelte"
+  import ScriptImporterModal from "./ScriptImporterModal.svelte"
   import { modal } from "../../../stores/editor"
   import { Modal } from "../../../constants/Modal"
 
@@ -10,7 +12,9 @@
     [Modal.Backups]: BackupsModal,
     [Modal.CreateProject]: CreateProjectModal,
     [Modal.LineFinder]: LineFinderModal,
-    [Modal.ItemFinder]: ItemFinderModal
+    [Modal.ItemFinder]: ItemFinderModal,
+    [Modal.TranslationKeys]: TranslationKeysModal,
+    [Modal.ScriptImporter]: ScriptImporterModal
   }
 
   $: component = components[$modal?.key]
