+++ app/javascript/src/constants/Modal.js
@@ -2,5 +2,7 @@ export const Modal = Object.freeze({
   Backups: "backups",
   CreateProject: "create-project",
   LineFinder: "line-finder",
-  ItemFinder: "item-finder"
+  ItemFinder: "item-finder",
+  TranslationKeys: "translation-keys",
+  ScriptImporter: "script-importer"
 })
